Move file RPCs to storage namespace
