+++ frontend/src/rpc/storage/files/index.ts
@@ -1,0 +1,11 @@
+// ================================================
+// WARNING: This file is automatically generated.
+// Do not modify it by hand. Your changes will be
+// overwritten the next time the generator runs.
+// ================================================
+
+import { rpcCall, FrontendFilesList1 } from '../../../shared/RpcModels';
+
+export const fetchList = (payload: any = null): Promise<FrontendFilesList1> => rpcCall('urn:storage:files:list:1', payload);
+export const fetchDelete = (payload: any = null): Promise<FrontendFilesList1> => rpcCall('urn:storage:files:delete:1', payload);
+export const fetchUpload = (payload: any = null): Promise<FrontendFilesList1> => rpcCall('urn:storage:files:upload:1', payload);
