+++ frontend/src/shared/RpcModels.tsx
@@ -88,37 +88,6 @@ export interface RoleItem {
   display: string;
   bit: number;
 }
-export interface AuthSessionTokens1 {
-  bearerToken: string;
-  rotationToken: string;
-  rotationExpires: any;
-}
-export interface AuthMicrosoftLoginData1 {
-  bearerToken: string;
-  defaultProvider: string;
-  username: string;
-  email: string;
-  backupEmail: string | null;
-  profilePicture: string | null;
-  credits: number | null;
-  rotationToken: string | null;
-  rotationExpires: any | null;
-}
-export interface FrontendVarsFfmpegVersion1 {
-  ffmpeg_version: string;
-}
-export interface FrontendVarsHostname1 {
-  hostname: string;
-}
-export interface FrontendVarsRepo1 {
-  repo: string;
-}
-export interface FrontendVarsVersion1 {
-  version: string;
-}
-export interface ViewDiscord1 {
-  content: string;
-}
 export interface FileItem {
   name: string;
   url: string;
@@ -152,6 +121,21 @@ export interface RouteItem {
   name: string;
   icon: string;
 }
+export interface FrontendVarsFfmpegVersion1 {
+  ffmpeg_version: string;
+}
+export interface FrontendVarsHostname1 {
+  hostname: string;
+}
+export interface FrontendVarsRepo1 {
+  repo: string;
+}
+export interface FrontendVarsVersion1 {
+  version: string;
+}
+export interface ViewDiscord1 {
+  content: string;
+}
 export interface FrontendUserProfileData1 {
   bearerToken: string;
   defaultProvider: string;
@@ -171,19 +155,41 @@ export interface FrontendUserSetDisplayName1 {
   bearerToken: string;
   displayName: string;
 }
-export interface ConfigItem {
-  key: string;
-  value: string;
+export interface AuthSessionTokens1 {
+  bearerToken: string;
+  rotationToken: string;
+  rotationExpires: any;
 }
-export interface SystemConfigDelete1 {
-  key: string;
+export interface AuthMicrosoftLoginData1 {
+  bearerToken: string;
+  defaultProvider: string;
+  username: string;
+  email: string;
+  backupEmail: string | null;
+  profilePicture: string | null;
+  credits: number | null;
+  rotationToken: string | null;
+  rotationExpires: any | null;
 }
-export interface SystemConfigList1 {
-  items: ConfigItem[];
+export interface SystemRouteDelete1 {
+  path: string;
 }
-export interface SystemConfigUpdate1 {
-  key: string;
-  value: string;
+export interface SystemRouteItem {
+  path: string;
+  name: string;
+  icon: string;
+  sequence: number;
+  requiredRoles: string[];
+}
+export interface SystemRouteUpdate1 {
+  path: string;
+  name: string;
+  icon: string;
+  sequence: number;
+  requiredRoles: string[];
+}
+export interface SystemRoutesList1 {
+  routes: SystemRouteItem[];
 }
 export interface SystemUserCreditsUpdate1 {
   userGuid: string;
@@ -213,26 +219,6 @@ export interface SystemUserRolesUpdate1 {
 export interface SystemUsersList1 {
   users: UserListItem[];
 }
-export interface SystemRouteDelete1 {
-  path: string;
-}
-export interface SystemRouteItem {
-  path: string;
-  name: string;
-  icon: string;
-  sequence: number;
-  requiredRoles: string[];
-}
-export interface SystemRouteUpdate1 {
-  path: string;
-  name: string;
-  icon: string;
-  sequence: number;
-  requiredRoles: string[];
-}
-export interface SystemRoutesList1 {
-  routes: SystemRouteItem[];
-}
 export interface SystemRoleDelete1 {
   name: string;
 }
@@ -251,6 +237,20 @@ export interface SystemRoleUpdate1 {
 }
 export interface SystemRolesList1 {
   roles: RoleItem[];
+}
+export interface ConfigItem {
+  key: string;
+  value: string;
+}
+export interface SystemConfigDelete1 {
+  key: string;
+}
+export interface SystemConfigList1 {
+  items: ConfigItem[];
+}
+export interface SystemConfigUpdate1 {
+  key: string;
+  value: string;
 }
 
 export async function rpcCall<T>(op: string, payload: any = null): Promise<T> {
