+++ rpc/metadata.json
@@ -69,18 +69,6 @@
       "capabilities": 0
     },
     {
-      "op": "urn:frontend:files:delete:1",
-      "capabilities": 0
-    },
-    {
-      "op": "urn:frontend:files:list:1",
-      "capabilities": 0
-    },
-    {
-      "op": "urn:frontend:files:upload:1",
-      "capabilities": 0
-    },
-    {
       "op": "urn:frontend:links:get_home:1",
       "capabilities": 0
     },
@@ -110,6 +98,18 @@
     },
     {
       "op": "urn:frontend:vars:get_version:1",
+      "capabilities": 0
+    },
+    {
+      "op": "urn:storage:files:delete:1",
+      "capabilities": 0
+    },
+    {
+      "op": "urn:storage:files:list:1",
+      "capabilities": 0
+    },
+    {
+      "op": "urn:storage:files:upload:1",
       "capabilities": 0
     },
     {
